refactor(worker): extract person_segment_sam3_common shared helpers (sc-8847)
F-045: the two cfg-exclusive SAM3 modules — the native-MLX
person_segment_sam3 (macOS) and the candle person_segment_sam3_candle
(off-Mac / backend-candle) — duplicated ~300 lines of pure, tensor-free
helpers VERBATIM: weight resolution, RGB->CHW normalization, and the
mask/association MATH that is the correctness core of Replace-Person and
the SCAIL-2 painter. A fix had to be applied twice or the platforms would
silently diverge.

Extract a backend-neutral person_segment_sam3_common compiled on BOTH
cfg lanes (same superset gate as scail2_masks:
any(macos, all(not(macos), backend-candle))) holding the shared helpers +
their tests ONCE. Each per-backend file keeps only its tensor/model/device
seam (input_tensor over Array vs Tensor+Device, the WEIGHTS cache, quant
parsing, the propagate driver, plus the MLX-only smart-select box/point
paths) and re-exports ensure_segmenter_weights + AllPersonMasks so
callers (media_jobs, video_jobs, segment_jobs, scail2_masks) are unchanged.

Helpers extracted: MODEL_FILE/TOKENIZER_FILE/SEG_REPO/INPUT_SIZE/
MASK_GRID/CONCEPT_PROMPT consts, BoxNorm, resolve_segmenter_weights,
ensure_segmenter_weights, normalize_chw, mask_box_containment,
select_object, mask_to_frame, mask_centroid_x, AllPersonMasks, and the
four pure-helper tests (normalize, containment, select_object,
mask_to_frame).

Byte-identity confirmed: every extracted function body is line-for-line
identical to the origin/main MLX originals (diff shows only the added
pub(crate) qualifier). select_object is generalized over a tiny
Sam3FrameOutput trait (obj_ids()/masks() accessors, impl'd per-backend on
each crate's structurally-identical VideoFrameOutput) so the identical
selection math runs on either backend without depending on either sam3
crate; the only diff vs the original is the generic signature + the
field->accessor swap, the scoring/tie-break math is unchanged.

rollup_mask_state is NOT touched — already consolidated into
media_jobs::mask_rollup_state by sc-8833; verified no stale duplicate
remains in the SAM3 modules.

No drift found: all eight helpers were byte-identical across the two
files except the candle copy's containment test lacked the MLX copy's
empty-mask divide-by-zero assertion; the shared test keeps the richer
MLX version (strictly more coverage).
